# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tax Filing Efficiency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Documents ────────────────────────────────────────────────────────────
  /clients/{clientId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxDocument"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a tax document and trigger AI extraction
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Uploaded document with extracted data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxDocument"
  /clients/{clientId}/documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a tax document
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── W-2 Data ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, firstName, lastName, email, filingStatus, state, taxYear, dependentsUnder17, otherDependents, createdAt, updatedAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        filingStatus:
          type: string
          enum: [single, married_filing_jointly, married_filing_separately, head_of_household, qualifying_widow]
        state:
          type: string
        taxYear:
          type: integer
        dependentsUnder17:
          type: integer
          description: Number of qualifying children under 17 with SSN (drives Child Tax Credit).
        otherDependents:
          type: integer
          description: Other qualifying dependents (drives the $500 Credit for Other Dependents).
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateClientBody:
      type: object
      required: [firstName, lastName, email, filingStatus, state, taxYear]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        filingStatus:
          type: string
          enum: [single, married_filing_jointly, married_filing_separately, head_of_household, qualifying_widow]
        state:
          type: string
        taxYear:
          type: integer
        dependentsUnder17:
          type: integer
        otherDependents:
          type: integer
        notes:
          type: ["string", "null"]
    UpdateClientBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        filingStatus:
          type: string
          enum: [single, married_filing_jointly, married_filing_separately, head_of_household, qualifying_widow]
        state:
          type: string
        taxYear:
          type: integer
        dependentsUnder17:
          type: integer
        otherDependents:
          type: integer
        notes:
          type: ["string", "null"]

    # ── TaxDocument ──────────────────────────────────────────────────────────
    TaxDocument:
      type: object
      required: [id, clientId, documentType, fileName, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        documentType:
          type: string
          enum: [w2, form_1099, form_1098, schedule_k1, other]
        fileName:
          type: string
        fileContent:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, extracted, failed]
        extractedText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UploadDocumentBody:
      type: object
      required: [documentType, fileName, fileContent]
      properties:
        documentType:
          type: string
          enum: [w2, form_1099, form_1098, schedule_k1, other]
        fileName:
          type: string
        fileContent:
          type: string
          description: Base64-encoded file content

    # ── W2Data ───────────────────────────────────────────────────────────────
...